refactor: move functions that read/write gcs and files to repositories/
