+++ src/builder/download-cursus-users.ts
@@ -1,19 +1,9 @@
 import { initializeApp, cert } from "firebase-admin/app";
 import { getStorage } from "firebase-admin/storage";
 import sub from "date-fns/sub";
-import { resolve } from "path";
-import dotenv from "dotenv";
-import { format, utcToZonedTime } from "date-fns-tz";
-import { getEnv } from "../utils/getEnv";
-import { writeFileSync } from "fs";
-import { CursusUser } from "../types/CursusUser";
-import {
-  getBeginAtList,
-  getEvaluationPoints,
-  getLevelBeginAtData,
-  getStudentStatusData,
-  isCurrentStudent,
-} from "../libs/process-cursus-users";
+import { format } from "date-fns-tz";
+import { downloadCursusUsersJson } from "../repositories/cloud-storage";
+import { saveContents } from "../repositories/local-file";
 
 const TIMEZONE = "Asia/Tokyo";
 const TIME_FORMAT = "yyyy-MM-dd HH:mm:ss";
@@ -34,20 +24,15 @@ initializeApp({
 });
 
 const downloadCursusUsers = async () => {
-  const bucket = getStorage().bucket();
   let now = new Date();
   for (let i = 0; i < MAX_DOWNLOAD_RETRY_COUNT; i++) {
     const filename = `v1/cursus_users/${format(now, "yyyy-MM-dd", {
       timeZone: TIMEZONE,
     })}.json`;
-    console.log(`try to download: ${filename}`);
     try {
-      const file = bucket.file(filename);
-
-      const [metadata] = await file.getMetadata();
-      const timeCreated = new Date(metadata.timeCreated);
-      const [buffer] = await file.download();
-      const cursusUsers: CursusUser[] = JSON.parse(buffer.toString());
+      const { cursusUsers, timeCreated } = await downloadCursusUsersJson(
+        filename
+      );
 
       const contents = aggregateContents(cursusUsers, timeCreated);
       await saveContents(contents);
@@ -97,11 +82,6 @@ const isStudent = (cursusUser: CursusUser) => {
     cursusUser.user.pool_year &&
     cursusUser.user.pool_year >= 2020
   );
-};
-
-const saveContents = async (contents: any) => {
-  const file = resolve(CURSUS_USERS_DIR, "contents.json");
-  await writeFileSync(file, JSON.stringify(contents));
 };
 
 (async () => await downloadCursusUsers())();
